LayoutTests/imported/w3c/web-platform-tests/navigation-timing/response-start-after-coop-bcg-switch.https-expected.txt


FAIL Navigation responseStart is not clamped to zero after a cross-origin COOP browsing context group switch assert_equals: The cross-origin navigation should not expose the redirect count expected 0 but got 2
